fix: updated with review comment
